move logic party to conn_party.go
